requirements.txt
fastapi==0.111.0
gradio==4.36.1
python-dotenv==1.0.1
Requests==2.32.3
